+++ R/descriptives.R
@@ -217,24 +217,3 @@ skew_kurtosis.default <-
       return(out[c(1, 3, 4, 6)])
     }
   }
-
-
-is_quiet <- function() {
-  isTRUE(getOption("usethis.quiet", default = FALSE))
-}
-
-#' @importFrom usethis ui_oops ui_done
-col_message <- function (..., col = 30, success = TRUE, verbose = !getOption("usethis.quiet", default = FALSE)){
-  if(verbose){
-    txt <- do.call(paste0, list(...))
-    # Check if this function is called from within an rmarkdown document.
-    # If that is the case, the colorized messages can cause knitting errors.
-    if(!any(grepl("rmarkdown", unlist(lapply(sys.calls(), `[[`, 1)), fixed = TRUE))){
-      if(success){
-        cli::cli_bullets(text = c("v" = txt))
-      } else {
-        cli::cli_bullets(text = c("x" = txt))
-      }
-    }
-  }
-}